Update an array through communication and request shutdown after
